# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CampusVoice API",
        "description": "API documentation for CampusVoice application",
        "contact": {
            "name": "CampusVoice Support",
            "email": "admin@campusvoice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "CampusVoice API Server"
        }
    ],
    "paths": {
...
        "/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Get all departments",
                "description": "Returns a list of all departments",
                "operationId": "getDepartments",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Department"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch departments"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Create a new department",
                "description": "Creates a new department and returns it",
                "operationId": "storeDepartment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name_ar",
                                    "name_fr"
                                ],
                                "properties": {
                                    "name_ar": {
                                        "type": "string",
                                        "example": "قسم علوم الحاسوب"
                                    },
                                    "name_fr": {
                                        "type": "string",
                                        "example": "Département d'Informatique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Department created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Department created successfully"
                                        },
                                        "department": {
                                            "$ref": "#/components/schemas/Department"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create department"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Department": {
                "title": "Department",
                "description": "Department model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the department",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name_ar": {
                        "title": "Name (Arabic)",
                        "description": "Name of the department in Arabic",
                        "type": "string",
                        "example": "قسم علوم الحاسوب"
                    },
                    "name_fr": {
                        "title": "Name (French)",
                        "description": "Name of the department in French",
                        "type": "string",
                        "example": "Département d'Informatique"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Creation date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Last update date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Department"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}